feat(endpoints): changes review @mattia
