+++ backend/app/repositories/search_profile_repository.py
@@ -3,17 +3,19 @@ from uuid import UUID
 
 from sqlalchemy.ext.asyncio import async_session
 from sqlalchemy.orm import selectinload
-from sqlmodel import select, and_, or_
+from sqlmodel import and_, or_, select
 
 from app.models.search_profile import SearchProfile
 from app.schemas.search_profile_schemas import SearchProfileUpdateRequest
 
 
 class SearchProfileRepository:
-    from sqlalchemy import or_, and_
+    from sqlalchemy import and_, or_
 
     @staticmethod
-    async def get_by_id(search_profile_id: UUID, current_user) -> SearchProfile | None:
+    async def get_by_id(
+        search_profile_id: UUID, current_user
+    ) -> SearchProfile | None:
         async with async_session() as session:
             result = await session.exec(
                 select(SearchProfile)
@@ -22,8 +24,9 @@ class SearchProfileRepository:
                     and_(
                         SearchProfile.id == search_profile_id,
                         or_(
-                            SearchProfile.is_public == True,
-                            SearchProfile.organization_id == current_user["organization_id"],
+                            SearchProfile.is_public,
+                            SearchProfile.organization_id
+                            == current_user["organization_id"],
                             SearchProfile.users.any(id=current_user["id"]),
                         ),
                     )
