+++ backend/app/schemas/match_schemas.py
@@ -1,38 +1,9 @@
-from datetime import datetime
 from typing import Literal
-from uuid import UUID
 
 from pydantic import BaseModel, Field
-from app.models.match import Match
 
 
 class MatchFeedbackRequest(BaseModel):
     comment: str
     reason: Literal["bad source", "false", "no good matching", "other"]
     ranking: int = Field()
-
-class ArticleOverviewItem(BaseModel):
-    id: UUID
-    title: str
-    url: str
-    author: str
-    published_at: datetime
-    language: str
-    category: str
-    summary: str
-    sorting_order: int
-
-    @classmethod
-    def from_entity(cls, match: Match) -> "ArticleOverviewItem":
-        article = match.article
-        return cls(
-            id=article.id,
-            title=article.title,
-            url=article.url,
-            author=article.author,
-            published_at=article.published_at,
-            language=article.language,
-            category=article.category,
-            summary=article.summary,
-            sorting_order=match.sorting_order,
-        )
